mini-services/rl-service/requirements.txt
# RL Service Dependencies
fastapi==0.109.0
uvicorn==0.27.0
numpy==1.26.3
pandas==2.2.0
gymnasium==0.29.1
stable-baselines3==2.2.1
torch==2.2.0
pydantic==2.6.0
pyyaml==6.0.1
tensorboard==2.15.1
aiohttp==3.9.1
python-multipart==0.0.6
